implemented a templete match detector method call to determine if a targeting reticle is present on the screen. if not it keeps pressing num0 until it appears. once it appears it advances to the next part of the loop. modified git ignore to ignore /tmp/ folder. added in template and mask for targeting reticle.
implemented a templete match detector method call to determine if a targeting reticle is present on the screen. if not it keeps pressing num0 until it appears. once it appears it advances to the next part of the loop. modified git ignore to ignore /tmp/ folder. added in template and mask for targeting reticle.